Give users/devs more insight into complex matchers
All of the validation matchers ultimately depend on allow_value to do
their work. When such a matcher fails, it can often be frustrating to
pin down the cause, because we don't know what the matcher is doing
under the hood.

This commit changes the error messages for all of the validation
matchers so that they list all the assertions they make and whether they
passed or failed.

It also implements #pretty_print for a bunch of internal classes so that
when developers use Pry to debug issues they can see into objects
without being overloaded with information they don't need.